Signed-off-by: Raghavan <ragavans85@gmail.com>
Least distance stage0.

This file is a sanity test module for the least distance algorithm.
The algorithm tries to identify the best match of a query (a subset)
within a reference (a superset).

stage0 denotes the stage of validation where
query is a non modified sub set of the reference