least_distance_stage1.py
This code will see if the mfcc + least distance can be used to identify
the location of a small part of an audio within a larger part of it.

The project name for this experiment stage1. stage0 did a comparison in the
mfcc domain, meaning a smaller mfcc subset was identified within a larger
mfcc superset. There in stage0, the location returned by the algorithm
was expected to be exactly the start and end of the query formed. The query
there was a set of mfcc's.

Here in stage1 the locations maynot be totally the same,
as the query, which is now a small duration of the audio maynot have its
start and end aligned with the hops made by the libsrosa module, resulting
in the reference audio and query audio giving slightly different mfcc's.

Hence the code will see if the algorithm returns the location within a
specified range around the location where the query was formed.

Even if the returned location exceeds the range, it might be a false
negative. Hence those locations have to be heard out manually to see if
the algorithm reurned a location (a part of the reference) which was
accoustically similar to the query.

As a way of documenting the results of the experiment,
pasting content from an email coversation :

Hi Ajit

Stage1 (in time domain). In stage0 we did a sanity test matching a subset within the superset, with the subset constructed from the superset. Here we take a 5 second reference. Divide it into 2 sec queries. 0-2, 1-3, 2-4 and 3-5.

I am printing when the match exceeds 2 frames. That is if the query starts at sample 1000 and ends at 2000, I will print a mismatch when the match begins at less than 980 or ends at more than 2020, with 10 samples in a frame.

I ran on 3 files. For one file, I got just a single mismatch. I heard the area, and it was a false negative. The query and the match sounded the same. On this file, I downsampled (to 8K) and ran the test. Got the same area as a mismatch. Heard the area and it sounded the same.

For another file, I got 25 mismatches. Heard a few. They sounded the same. What I observed is when the query includes silence, the match is much shorter. Say the query is 2 secs, with 1 sec silence. It matches with the 1 sec audio in the reference. I guess this is a good thing. I downsampled to 8K. Got just 6 mismatches. They were all false negatives. Most of them being queries with some silent parts getting matched with the audio part.

For another file there were about 10 mismatches. But a few were such that the start and the end of the query was the same sample. The 44k samples (2 sec * 22k sample rate) had been aligned with just one sample. I heard the query. It was a constant string (veenai) sound. I will look into this, or maybe since it is just the string part, maybe we don't need to pay attention. When I downsampled to 8K, I did not get any mismatch on this dataset.

So, my general observation is - it is working.

------------ ANOTHER EMAIL CONVERSATION ----------------

Just an update. Was just analysing the best matches and the 6 mismatches.

In cases when the query is a recitation, the matches are good. When the query is a silence or a background swara alone, the results are not good. The match does not ever overlap with the query. If the query was 2-4 sec, the match is at 4.8 sec (and as said before )