This code sees if the mfcc's along with least distance algorithm can be used to match a small part of an audio within a large part, when the small part is modified synthetically (using inbuilt librosa functions that do time stretch and pitch shift).
This is stage2. Stage1 was where the comparision was done without the
synthetic mofidifcations.
That is a small duration of the audio is used as query and the larger
duration is used as reference.

In stage0 the comparision was done at the mfcc domain, meaning a small
subset of mfcc's was queried against a superset of mfcc's

If PITCH_SHIFT is on, the pitch will be shifted by a factor of
LIBROSA_SEMITONE_STEP. A negative LIBROSA_SEMITONE_STEP will mean that the
pitch has been lowered, a positive value will mean that the pitch has been
made higher. If TIME_STRETCH is on, for a positive LIBROSA_TIME_STRETCH
value, the audio will be made faster and correspondingly altered for a
negative LIBROSA_TIME_STRETCH value.

The reference will be formed by combining NUM_QUERIES_IN_REFERENCE chunks.
The query_directory is the folder that holds the sishya app audio clips
(quaters of a sentence). There are 76 clips in the current directory.
If NUM_QUERIES_IN_REFERENCE is 10, then every 10 of the 76 would be combined
to form a reference. The last 6 would form the final reference.
Then every chunk within the reference is taken for modification.
The modified audio is fed as the query to the algorithm. The algorithm is
expected to match the query within the reference at its actual location.
A match is when the algorithm returned start and end lie within
MATCH_THRESHOLD.